+++ week3/d5-mvc-jsp/d5_mvc_jsp_quiz/src/main/java/com/example/d5_mvc_jsp_quiz/domain/Quiz.java
@@ -10,7 +10,7 @@ import java.util.List;
 @AllArgsConstructor
 @ToString
 public class Quiz {
-    private int id;
-    private String category;
-    private List<Question> questionList;
+  private Long id;
+  private String category;
+  private List<Question> questionList;
 }
