+++ week3/d5-mvc-jsp/d5_mvc_jsp_quiz/src/main/java/com/example/d5_mvc_jsp_quiz/domain/User.java
@@ -8,11 +8,11 @@ import lombok.*;
 @AllArgsConstructor
 @ToString
 public class User {
-    private int id;
-    private String username;
-    private String password;
-    private String email;
-    private int role; // 0-user, 1-admin
-    private String firstName;
-    private String lastName;
+  private Long id;
+  private String username;
+  private String password;
+  private String email;
+  private int role; // 0-user, 1-admin
+  private String firstName;
+  private String lastName;
 }
